requirements.txt
-e git+https://github.com/Rohitrajsaji/AgenticAI.git@3cac8263cec2be2b1b6a9c0ec13c6544cc95aee7#egg=agentic_ai
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
duckduckgo_search==6.4.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.4.0
numpy==2.3.2
openai==1.107.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
tavily-python==0.7.11
tenacity==8.5.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
